backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
pyjwt==2.9.0
email-validator==2.3.0
python-multipart==0.0.9
requests==2.32.3
reportlab==4.2.0
